cleanup actions, move useActiveTimeSeriesQueries to runtime/time-series-queries.ts
